tuning

+++ collect_state_data.py
@@ -98,7 +98,7 @@ class StackingObsParser(object):
 def main(args):
     config = dict(
         env_id="BulletStack-v2",
-        num_workers=4,
+        num_workers=64,
         algo="ppo",
         name="debug",  # pyramid_base_9obj_3goal_roll
         total_timesteps=int(2e8),
@@ -143,7 +143,7 @@ def main(args):
     )
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
     env = make_vec_env(config["env_id"], config["num_workers"], device, log_dir=None,
-                       training=False, **config["create_env_kwargs"])
+                       training=False, reset_when_done=True, **config["create_env_kwargs"]) # TODO: note reset when done
     obs_parser = config["obs_parser"]
     policy = AttentionDiscretePolicy(obs_parser, env.action_space.shape[0], **config["policy"])
     policy.to(device)
@@ -155,11 +155,13 @@ def main(args):
     # if os.path.exists("tmp"):
     #     shutil.rmtree("tmp")
     # os.makedirs("tmp")
-    # create_generalize_task(env, image_processor, 1000, use_patch_feat=False, use_raw_img=False if use_pretrain else True)
+    # create_generalize_task(env, image_processor, 1000, use_patch_feat=False, use_raw_img=False if use_pretrain else True, shape="Y")
     # create_canonical_view(env, image_processor, use_patch_feat=False, use_raw_img=True)
-    # collect_data(env, policy, image_processor, 10_000, balance=False, round_idx=args.round_idx, use_patch_feat=False, use_raw_img=False if use_pretrain else True)
+    n_traj = min(len(env.get_attr("offline_datasets")[0]), 10_000)
+    collect_data(env, policy, image_processor, n_traj, balance=False, round_idx=args.round_idx, use_patch_feat=False, use_raw_img=False if use_pretrain else True)
+    # follow_expand_data(env, obs_parser, image_processor, n_traj, args.round_idx)
     # create test tasks
-    prepare_visual_base_data(env, policy, image_processor, "collect_data_last_step.pkl", use_patch_feat=False, use_raw_img=False if use_pretrain else True)
+    # prepare_visual_base_data(env, policy, image_processor, "collect_data_last_step.pkl", use_patch_feat=False, use_raw_img=False if use_pretrain else True)
 
     return
     if os.path.exists("tmp"):
@@ -184,11 +186,15 @@ def collect_data(env, policy, image_processor, desired_num, balance, round_idx, 
     dataset_fname = "distill_dataset_new_stacking_%sexpand%d.pkl" % ("raw_" if use_raw_img else "", round_idx)
     task_fname = "distill_tasks_new_%sexpand%d%s.pkl" % ("raw_" if use_raw_img else "", round_idx, "_balance" if balance else "")
     if os.path.exists(dataset_fname):
-        os.remove(dataset_fname)
+        ans = input(f'{dataset_fname} exists, remove? [Y|n]')
+        if ans == "Y":
+            os.remove(dataset_fname)
+        else:
+            print("Not removed")
+            return
     assert env.get_attr("use_expand_goal_prob")[0] == 1
     device = image_processor.device
     n_object = 6
-    im_dataset = [dict(obs=[], action=[], boundary=[]) for _ in range(n_object)]
     task_arrays = [[] for _ in range(n_object)]
     traj_buffer = [dict(img=[], state=[], action=[]) for _ in range(env.num_envs)]
     obs = env.reset()
@@ -199,14 +205,14 @@ def collect_data(env, policy, image_processor, desired_num, balance, round_idx, 
     n_data = [0 for _ in range(n_object)]
     while sum([len(task_arrays[i]) for i in range(n_object)]) < desired_num and sum(n_data) < 6 * desired_num:
         with torch.no_grad():
-            _, action, _, _ = policy.act(obs)
+            _, action, _, _ = policy.act(obs, deterministic=True)
         obs, reward, done, info = env.step(action)
         for i in range(env.num_envs):
             traj_buffer[i]["action"].append(action[i].detach().cpu().numpy())
             traj_buffer[i]["state"].append(obs[i].detach().cpu().numpy())
             traj_buffer[i]["img"].append(info[i]["img"])
             if done[i]:
-                if info[i]["is_success"] and reward[i] == 1 and len(traj_buffer[i]["action"]) < 20:
+                if info[i]["is_success"] and reward[i] == 1 and len(traj_buffer[i]["action"]) < 10:
                     traj_buffer[i]["state"][-1] = info[i]["terminal_observation"]
                     # Bridge to pixel env observation, relabel
                     # TODO
@@ -274,13 +280,97 @@ def collect_data(env, policy, image_processor, desired_num, balance, round_idx, 
         pickle.dump(all_tasks, f)
 
 
-def create_generalize_task(env, image_processor, desired_num, use_patch_feat, use_raw_img):
-    task_fname = "test_tasks_%s.pkl" % ("raw" if use_raw_img else "")
+# only prefix is recorded, do not use this function
+def follow_expand_data(env, policy, image_processor, desired_num, round_idx):
+    dataset_fname = f"distill_dataset_new_stacking_raw_expand{round_idx}.pkl" 
+    task_fname = f"distill_tasks_new_raw_expand{round_idx}.pkl"
+    if os.path.exists(dataset_fname):
+        ans = input(f'{dataset_fname} exists, remove? [Y|n]')
+        if ans == "Y":
+            os.remove(dataset_fname)
+        else:
+            print("Not removed")
+            return
+    assert env.get_attr("use_expand_goal_prob")[0] == 1
+    n_object = 6
+    task_arrays = [[] for _ in range(n_object)]
+    traj_buffer = [dict(img=[], state=[], action=[]) for _ in range(env.num_envs)]
+    actions_to_follow = [None for _ in range(env.num_envs)]
+    obs = env.reset()
+    imgs = env.get_images()
+    for i in range(env.num_envs):
+        traj_buffer[i]["state"].append(obs[i].detach().cpu().numpy())
+        traj_buffer[i]["img"].append(imgs[i].astype(np.uint8))
+        actions_to_follow[i] = env.get_attr("expand_action_seq", indices=i)[0]
+    n_data = [0 for _ in range(n_object)]
+    while sum([len(task_arrays[i]) for i in range(n_object)]) < desired_num and sum(n_data) < 6 * desired_num:
+        with torch.no_grad():
+            _, action, _, _ = policy.act(obs)
+        for i in range(env.num_envs):
+            if len(actions_to_follow[i]):
+                action[i] = torch.from_numpy(actions_to_follow[i][0]).to(obs.device)
+                actions_to_follow[i] = actions_to_follow[i][1:]
+        obs, reward, done, info = env.step(action)
+        for i in range(env.num_envs):
+            traj_buffer[i]["action"].append(action[i].detach().cpu().numpy())
+            traj_buffer[i]["state"].append(obs[i].detach().cpu().numpy())
+            traj_buffer[i]["img"].append(info[i]["img"])
+            if done[i]:
+                # TODO: if success
+                # TODO: modify to auto reset to prevent infinite rollout
+
+                # print("should done", done[i])
+                # print("obs", obs[i], "goal", env.get_attr("goal", indices=i)[0])
+                traj_states = np.stack(traj_buffer[i]["state"], axis=0)
+                traj_images = np.stack(traj_buffer[i]["img"], axis=0)
+                traj_features = image_processor.mvp_process_image(
+                    traj_images, use_patch_feat=False, use_raw_img=True).detach().cpu().numpy()
+                robot_obs, objects_obs = policy.obs_parser.bridge(traj_states)
+                # print("last objects obs", objects_obs[-1])
+                goal_idx = len(traj_buffer[i]["action"])
+                goal_feature = traj_features[goal_idx:goal_idx + 1]
+                goal_state = objects_obs[goal_idx:goal_idx + 1]     
+                new_obs = np.concatenate([
+                    traj_features[:goal_idx],
+                    np.tile(goal_feature, (goal_idx, 1)),
+                    robot_obs[:goal_idx],
+                    objects_obs[:goal_idx],
+                    np.tile(goal_state, (goal_idx, 1))
+                ], axis=-1).astype(np.float32)
+                n_to_move = len(set([int(traj_buffer[i]["action"][j][0]) for j in range(goal_idx)]))
+                n_data[n_to_move - 1] += new_obs.shape[0]
+                task_array = np.concatenate([
+                    robot_obs[0], objects_obs[0], objects_obs[goal_idx], traj_features[goal_idx]
+                ])
+                task_arrays[n_to_move - 1].append(task_array)
+                with open(dataset_fname, "ab") as f:
+                    pickle.dump({"obs": new_obs, "action": np.stack(traj_buffer[i]["action"][:goal_idx], axis=0).astype(np.float32)}, f)
+                print(sum([len(task_arrays[i]) for i in range(n_object)]), sum(n_data))
+                obs[i] = torch.from_numpy(env.env_method("reset", indices=i)[0]).float().to(obs.device)
+                actions_to_follow[i] = env.get_attr("expand_action_seq", indices=i)[0]
+                reset_img = env.env_method("render", indices=i)[0]
+                traj_buffer[i] = dict(
+                    img=[reset_img.astype(np.uint8)], state=[obs[i].detach().cpu().numpy()], action=[]
+                )
+    for i in range(n_object):
+        print("N to move", i + 1, "count", len(task_arrays[i]))
+    all_tasks = []
+    for i in range(n_object):
+        all_tasks.extend(task_arrays[i])
+    all_tasks = np.stack(all_tasks, axis=0)
+    # with open("distill_dataset_stacking_raw_expand%d%s.pkl" % (round_idx, "_balance" if balance else ""), "wb") as f:
+    #     pickle.dump(im_dataset, f)
+    with open(task_fname, "wb") as f:
+        pickle.dump(all_tasks, f)
+
+
+def create_generalize_task(env, image_processor, desired_num, use_patch_feat, use_raw_img, shape="3T"):
+    task_fname = "test_tasks_%s_%s.pkl" % ("raw" if use_raw_img else "", shape)
     all_task_arrays = []
     all_state_task_trajs = []
     count = 0
     while count < desired_num:
-        tasks = env.env_method("create_generalize_task")
+        tasks = env.env_method("create_generalize_task", shape=shape)
         robot_obs, init_states, goal_states, goal_images = map(lambda x: np.array(x), zip(*tasks))
         goal_feat = image_processor.mvp_process_image(
             goal_images, use_patch_feat=use_patch_feat, use_raw_img=use_raw_img
@@ -316,7 +406,7 @@ def create_generalize_task(env, image_processor, desired_num, use_patch_feat, us
     all_task_arrays = np.concatenate(all_task_arrays, axis=0)
     with open(task_fname, "wb") as f:
         pickle.dump(all_task_arrays, f)
-    with open("test_state_tasks.pkl", "wb") as f:
+    with open(f"test_state_tasks_{shape}.pkl", "wb") as f:
         pickle.dump({"expansion": all_state_task_trajs}, f)
 
 def create_canonical_view(env, image_processor, use_patch_feat, use_raw_img):
